Non-Functional
Currently trying to figure out a way to scan the text, a problem comes where the website did not use any of the By tools (such as IDs) which would be useful (byCSS wouldn't be particularly useful for differentiating between terms) for categorizing the elements.

Current idea is to search for the header corresponding to the course code, going back to the main header where the text for the corresponding information is also stored, and store that information. This way we are indexing and storing information in a general way.